requirements.txt
streamlit
pinecone
langchain-classic
langchain-openai
langchain-community
langchain-pinecone
langchain-text-splitters
python-dotenv
pypdf
